+++ frontend/src/components/Spinner/LoadingSpinner.js
@@ -1,0 +1,12 @@
+import CircularProgress from "@material-ui/core/CircularProgress";
+import React from "react";
+
+
+export default function LoadingSpinner() {
+
+  return (
+      <div>
+        <CircularProgress/>
+      </div>
+  );
+}
